Extract `FileBuilder` methods into classes
