+++ lib/annotate_rb/model_annotator.rb
@@ -27,5 +27,6 @@ module AnnotateRb
     autoload :FileComponents, "annotate_rb/model_annotator/file_components"
     autoload :ProjectAnnotator, "annotate_rb/model_annotator/project_annotator"
     autoload :ProjectAnnotationRemover, "annotate_rb/model_annotator/project_annotation_remover"
+    autoload :AnnotatedFile, "annotate_rb/model_annotator/annotated_file"
   end
 end
