+++ lib/annotate_rb/model_annotator/annotated_file.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module AnnotateRb
+  module ModelAnnotator
+    module AnnotatedFile
+      autoload :Generator, "annotate_rb/model_annotator/annotated_file/generator"
+      autoload :Updater, "annotate_rb/model_annotator/annotated_file/updater"
+    end
+  end
+end
